requirements.txt
beautifulsoup4==4.8.1
certifi==2019.11.28
chardet==3.0.4
Django==2.2.7
django-appconf==1.0.3
django-crispy-forms==1.8.0
django-extensions==2.2.6
django-livereload-server==0.3.2
djangorestframework==3.10.3
httplib2==0.17.0
idna==2.8
pkg-resources==0.0.0
psycopg2-binary==2.8.4
pytz==2019.3
rcssmin==1.0.6
requests==2.22.0
rjsmin==1.1.0
shortuuid==0.5.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
tornado==6.0.3
urllib3==1.25.8
